+++ src/Results/ProxyFeedbackInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace paslandau\QueryScraper\Results;
+
+
+interface ProxyFeedbackInterface {
+
+    /**
+     * @return string
+     */
+    public function getProxyResult();
+}