+++ changelog.md
@@ -1,3 +1,3 @@
 #dev-master
 
-- Initial commit
+- added a request generator instead of a simple array for Guzzle requests so that further requests can be appended while running. Needed because `$event->retry()` could not be used due to URL limitations during redirects